+++ plugins/render_delegate/renderer_plugin.cpp
@@ -36,6 +36,7 @@
 #include "render_delegate.h"
 
 #include <constant_strings.h>
+#include <pxr/base/tf/getEnv.h>
 
 PXR_NAMESPACE_OPEN_SCOPE
 
@@ -60,6 +61,10 @@ HdRenderDelegate* HdArnoldRendererPlugin::CreateRenderDelegate(const HdRenderSet
           houdiniRenderer->UncheckedGet<std::string>() == str::t_husk.GetString()))) {
         context = str::t_husk;
         isBatch = true;
+    }
+    if (TfGetenv("ARNOLD_FORCE_ABORT_ON_LICENSE_FAIL", "0") != "0" && !AiLicenseIsAvailable()) {
+        AiMsgError("Arborting: ARNOLD_FORCE_ABORT_ON_LICENSE_FAIL is set and Arnold license not found");
+        return nullptr;
     }
     auto* delegate = new HdArnoldRenderDelegate(isBatch, context, nullptr);
     for (const auto& setting : settingsMap) {